// src/types/partials/ExplanationOfBenefitItem.ts
/* eslint-disable no-unused-vars */
/* eslint-disable no-trailing-spaces */
// This file is auto-generated by generate_types so do not edit manually

import { TExtension } from '../partials/Extension';
import { TInt } from '../simpleTypes/Int';
import { TCodeableConcept } from '../partials/CodeableConcept';
import { TDate } from '../simpleTypes/Date';
import { TPeriod } from '../partials/Period';
import { TAddress } from '../partials/Address';
import { TReference } from '../partials/Reference';
import { TQuantity } from '../partials/Quantity';
import { TMoney } from '../partials/Money';
import { TDecimal } from '../simpleTypes/Decimal';
import { TExplanationOfBenefitAdjudication } from '../partials/ExplanationOfBenefitAdjudication';
import { TExplanationOfBenefitDetail } from '../partials/ExplanationOfBenefitDetail';

export type TExplanationOfBenefitItem = {
    id?: String;
    extension?: TExtension[];
    modifierExtension?: TExtension[];
    sequence: TInt;
    careTeamSequence?: TInt[];
    diagnosisSequence?: TInt[];
    procedureSequence?: TInt[];
    informationSequence?: TInt[];
    revenue?: TCodeableConcept;
    category?: TCodeableConcept;
    productOrService: TCodeableConcept;
    modifier?: TCodeableConcept[];
    programCode?: TCodeableConcept[];
    servicedDate?: TDate;
    servicedPeriod?: TPeriod;
    locationCodeableConcept?: TCodeableConcept;
    locationAddress?: TAddress;
    locationReference?: TReference;
    quantity?: TQuantity;
    unitPrice?: TMoney;
    factor?: TDecimal;
    net?: TMoney;
    udi?: TReference[];
    bodySite?: TCodeableConcept;
    subSite?: TCodeableConcept[];
    encounter?: TReference[];
    noteNumber?: TInt[];
    adjudication?: TExplanationOfBenefitAdjudication[];
    detail?: TExplanationOfBenefitDetail[];
};